+++ js/utils/get-random-positive-integer.js
@@ -28,3 +28,17 @@ function getRandomPositiveInteger (a, b) {
 }
 
 getRandomPositiveInteger();
+
+const getRandom = (min, max) => {
+  if (min < 0 || max < 0) {
+    return -1;
+  }
+
+  if (min > max) {
+    [min, max] = [max, min];
+  }
+
+  return Math.round(Math.random() * (max - min)) + min;
+};
+
+getRandom (1, 9);
